Remove the only place in parser where we accidentaly share same structure for two nodes
Summary: In dozens of places in parser where we need two missing nodes at the same offset, we call make_missing twice. This is the only one where we re-use the value. It is functionally equivalent, but gets annoying when we get down to comparing structures on a byte by byte level (which we will). Instead of constantly special casing few test cases, just changing it to behave like rest of the parser.

Reviewed By: kmeht

Differential Revision: D14574688

fbshipit-source-id: bd7e7b562654bd650758a5f43cdefb0d3bc72443