Remove Little_magic_cache from next_token
Summary:
I've noticed that removing this cache changes some outputs, which makes it not a very good cache (it is supposed to affect only performance, not correctness). It got broken in D8716143 where a `~tokenizer` param was added - if we call `next_token` twice at same position, but with different tokenizer, we'll get the same result (when a different one is expected).

We could fix it by providing two versions of `next_token` that have two separate caches, but this cache actually doesn't seem to do anything - both according to my measurements, and the diff that added it (D6245332): according to test plan, the performance was "(same as before (unfortunately))". Not sure why it was committed then.

Removing it reveals another bug: it is possible to trip `assert_xhp_body_token` assertion. This is the case of two calls to `next_token` at same position returning two different values due to switch between XHP and non-XHP context. I fix it by passing the already lexed token into the function.

Reviewed By: kmeht

Differential Revision: D14574614

fbshipit-source-id: 920f6d8394273c473f362c97125e7122cb215602