Remove implementation of assignment by reference from emitter
Summary:
The previous diff disabled this support in the parser. This diff assumes that
the parser does not allow references on RHS of assignments, turns this case
into an invariant check and simplifies the rest of the code.

`emit_lval_op()` also contained a special support for evaluation of
`list(...) = EXPR` by reference, which would leave the result of `EXPR`
evaluated by reference on the stack. This support was removed as well for
two reasons:
- parser does not allow binary expressions in write context, so
  `&(list() = ...)` is a parse error and there are no other ways to enter
  `emit_expr ~need_ref:true` after `=&` was banned and by-ref annotations
  enforcement was implemented
- it's inconsistent with `$foo = EXPR`, which does not have the same behavior
  when evaluated by reference

Reviewed By: oulgen

Differential Revision: D14627133

fbshipit-source-id: 4816e12d84894d50bb1ea87cdabab9010c799634