Convert emit_switch() to not use unnamed locals
Summary:
Switch has two different behaviors depending on whether its input is a local.
The "behavior" with local already did not utilize provided unnamed local. The
generic case utilized the unnamed local in a way that could be simply replaced
with keeping the evaluated input on the top of the stack.

This allows removal of with_temp_local and stash_in_local utilities and kills
last undesirable direct usage of Unwind opcode.

Reviewed By: billf

Differential Revision: D14655300

fbshipit-source-id: f8ef391ebc264c2db7aa47fe89fa7aa416f7793d