Check dumpTCAnnotation in dumpIREnabled and enable TC annotations for all Optimize translations
Summary:
When Eval.DumpIR was enabled, dumpIREnabled was returning true for all
translations, regardless whether we were really going to store the TC IR
annotations.  This incurred some significant JIT overhead, which impacted server
warmup.  This diff changes dumpIREnabled to check dumpTCAnnotation in case
DumpIR is enabled.  This significantly reduces the overhead of DumpIR, which
allows collecting TC annotations for the Optimize translations of all functions
(not just the ones marked as "hot"). This change is also included here.

Reviewed By: paulbiss

Differential Revision: D14633130

fbshipit-source-id: 4b42b47916cbfc23f2613f91e9da010135fe8746