Replace md5 with sha1 hashing in the hhbc and hhbbc repos
Summary:
HHVM uses MD5 hashes to keep track of the content of files in the repo, while Watchman delivers SHA1 hashes of files. And I've been told that when using the Eden filesystem, we'll very much want the hashes provided by Watchman instead of hashing files ourselves. If we want HHVM to use and trust Watchman's hashes internally, we need to use SHA1 for compatibility.

This diff was *mostly* accomplished with case-sensitive find-and-replace. The only significant differences are caused by the fact that an MD5 hash is 16 bytes, while a SHA1 hash is 20 bytes. This also means that any 32-byte hex strings should become 40-byte hex strings.

Reviewed By: paulbiss

Differential Revision: D13820401

fbshipit-source-id: b6ad690ea78b22815e4f876e3edbc80f5ce4f1ba