Fix incorrect parser error message
Summary:
Ever since the function `parse_xhp_close` was added in {D3431163}, in all its error branches, we use error1039 (A closing XHP tag is expected here) except the last, where we use a completely irrelevant and nonsensical message, error1026 (An opening square bracket is expected here).

I suspect that this was a copy/paste error and error1039 was intended.

Reviewed By: kmeht

Differential Revision: D14655451

fbshipit-source-id: bcee683d0a9604750bf79aa6d227a52bc262a499