Don't attempt to lex docstrings in XHP bodies
Summary:
Before {D8480896}, when we encountered a `<<<` token in an XHP body, the FFP would crash. In order to deal with this, D8480896 had the lexer attempt to scan a docstring literal instead. Docstring literals are not legal in an XHP body, and even if they were, this change did not result in parsing them correctly--parsing a docstring literal here causes the parser go off the rails in a way that I don't fully understand.

It is no longer the case that without D8480896, the parser crashes when it encounters a  `<<<` token in an XHP body, so this change reverts D8480896.

This change also improves the error message we get in this case. If we simply revert D8480896, we get the error `An expression is expected here`, which does not make sense--this error message was selected for the case where we are not in an XHP body context, and we have chosen to interpret a `<` token as a less-than operator (and thus expect to see an expression following it). We should emit a different error message in an XHP body context, where it is not reasonable to interpret `<` as a less-than operator. Before switching to the FFP, we produced a [much more sensible error](diffusion/FBS/browse/master/fbcode/hphp/hack/test/typecheck/heredoc_token_inside_xhp.php.exp;15c8f2a950f1bf626ea39e8cce26e08b7ba82a7d), `Stray < in xhp`. I think the error `A name is expected here` is a reasonable choice (since the `<` token can only legally be the beginning of an XHP expression, so we expect to see an XHP class name).

Reviewed By: kmeht

Differential Revision: D14634295

fbshipit-source-id: 10db1153ab5211da16abdec53beac846c7ad15cd