Watchman should throw if it has restarted
Summary:
Sometimes Flow/Hack lose their connection to Watchman. Hack & Flow will notice that the connection has died and will try to reconnect. This is all well and good. However, sometimes Watchman restarted. This may happen when a new Watchman is deployed or when the service dies for some reason.  When Watchman restarts, some file changes may be missed. In these cases, Flow & Hack need to restart too, just to be safe.

{D4143924} attempted to do this too a few years ago. Unfortunately, I'm not too sure if that worked. The `subscribe` query's response doesn't include the `is_fresh_instance` field. When `empty_on_fresh_instance` is `false`, `subscribe` will immediately send a huge response with `is_fresh_instance` set, but when `empty_on_fresh_instance` is `true`, nothing is pushed.

Reviewed By: Wilfred

Differential Revision: D14619612

fbshipit-source-id: 61b2ed6ecb6c4c008b93af437db122325de99751