Fix line numbers for SetL/PopC in for loops
Summary:
The PopC associated with the first and third expressions in a for loop
was given the line number of the end of the loop. This meant that when
optimizing SetL/PopC to PopL in hhbbc it was impossible to accurately
preserve line numbers. We had some heuristics that did well enough to
keep our code coverage tests passing in both non-repo and repo mode;
but I now have some optimizations that break them because there are no
instructions that we can mark with the original line number of the PopC.

I'm assuming this was just copying some strangeness from the original
hphpc emitter - so lets just fix it to do the sensible thing; the SetL
and the PopC should have the same line number.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D14683678

fbshipit-source-id: 582100302089caafa608f34241ff4a8e144efbee