Properly handle exceptions from calls at the end of try block
Summary:
Use the call offset rather than the offset after the call when looking up
exception handler. Otherwise an exception coming from a call at the end of try
block would not get properly caught.

Use the offset after the call only for the purpose of discarding stack temps.

It's a bit gross, but the whole logic will be simplified once faults are gone.

Reviewed By: ricklavoie

Differential Revision: D14675828

fbshipit-source-id: f3ebf6707366db74891ae8cc719d374b40d3efa6