IncDec Warn/Fatal on invalid types
Summary:
We wish to forbid Inc and Dec on non numeric types.  This option gives us a migration pathway to achieve this.

HHBBC apparently doesn't eliminate incdec's because they have side-effects that modify locals.  This doesn't change any of the HHBBC optimizations, that is left for after the cleanup is complete, and we are removing the option.

Reviewed By: billf

Differential Revision: D14508934

fbshipit-source-id: 63bb3853f7d56bf532e7715db87c6cb5e51dc81f