Fix instruction regression if skip type refinement for ClsMeth type
Summary:
Add `IsCompatibleClsMethType` runtime option. Turn it off by default to avoid instruction regression.
This flag is used to skip ClsMeth type refinement in order to make ClsMeth type be compatible with BArr/BVArr/BVec types.

If we turn on IsCompatibleClsMethType, `EmitClsMethPointers` also need to be true.
When codemod is done, this flag and the logic of skipping clsmeth type refinement can  be removed together.

https://fb.workplace.com/groups/1440831316224386/permalink/2088930781414433/

Reviewed By: markw65

Differential Revision: D14670570

fbshipit-source-id: 05eaf4cd3869e33c22145a88df2fa47fcb2a150c