Fall back to eager solving when narrowing fails
Summary: Narrowing of a type variable may fail if the widened type that is added as an upper bound to the type variable is incompatible with its existing lower bounds. In that case, we don't really want to see the subtyping errors arising from the checks implied by transitivity because they are hard to explain.  Instead, we can fall back to eager solving and push the type variable to its lower bound.  I expect this to produce better error messages because we still perform case analysis on types in many places.  See the new test case below for an example.

Reviewed By: andrewjkennedy

Differential Revision: D14539948

fbshipit-source-id: a4023c22ed26b2d85fd8c195c9d64524680dcfc4