Fix type checking of nullsafe shape accesses
Summary:
At the moment, expressions like `$s['x'] ?? $y` where the type of `$s` is a shape that doesn't contain a field named `'x'` (e.g., `shape()`) fail to type check, even though they should because (a) by applying the subsumption rule we should be able to treat `$s` as a shape in which `'x'` is declared as an optional field (e.g., we should be able to push the type of `$s` up to `shape(?'x' => nothing)`), and (b) the expression is actually safe because of the special run time behavior of the coalesce operator `??` (`$s['x']` throws if `'x'` doesn't exist in `$s` but `$s['x'] ?? $y` is evaluated as if `$s['x']` returned `null`).

It's easy to see that the expression `$s['x'] ?? $y` is semantically equivalent to `Shapes::idx($s, 'x') ?? $y`, which gives an idea for type-checking `$s['x']` appearing on the left of the coalesce operator: just treat it as `Shapes::idx($s, 'x')`.  Note that if the type of `$s` is known to be a shape that contains a required field `'x'` we will infer a type for `$s['x'] ?? $y` that is a union of the type of `'x'` and the type of `$y`, even though we could be more precise and not include the type of `$y` into the results (and provide a lint warning about a redundant coalesce operator) -- but we already do this, so the types are not getting less precise with this change.

Reviewed By: nbenton

Differential Revision: D14665376

fbshipit-source-id: de5055a38588f4763fc3daa8ec33a33c87151522