New inference: don't reject array indexing where nothing is inferred
Summary: D14520917 was a bit too strict in rejecting array indexing. Often we end up inferring `nothing` for the key type of an array, just as the legacy checker inferred `Tunresolved[]`. In these cases, we should not produce an error on indexing the array, as it's likely that the inferred type came about because of interaction with untyped code. (In any case, as discussed elsewhere, this isn't really a *type system* error, but is more like a lint warning that indexing might fail at runtime.)

Reviewed By: manzyuk

Differential Revision: D14725637

fbshipit-source-id: 139cb3978d9c11b144afbff6aba3cbb04c5cc647