Move responsibility for popping object to FPushObjMethod*'s helpers
Summary:
Moving trivial parts of the calling convention to separate cleanup diffs to
make it easier to find the regression.

Do not pass object explicitly to fPushObjMethodImpl(), let it grab it from the
stack. Kills unnecessary try/catch. There's still a memory leak for name, but
that's for another diff.

Similarly, move responsibility for popping the null to fPushNullObjMethod().

Reviewed By: alexeyt

Differential Revision: D14691644

fbshipit-source-id: bc15de954e0963075e00530996a38cce4c6e0a77