Fix opam2/arm64 build (#8474)
Summary:
The current ocaml used to build hhvm is from opam2, which is lacking
the patches/ocaml.patch patch. This PR updates the opam switch creation
so that ocaml is patched before it is built.

The old ocaml is still present to have a simple PR. Next PR will remove
the old ocaml from third-party and the toplevel ocaml-patch
Pull Request resolved: https://github.com/facebook/hhvm/pull/8474

Reviewed By: vassilmladenov

Differential Revision: D14742570

Pulled By: vsiles

fbshipit-source-id: 80b6b9f37bfc7c059ec1db61aa48d33f42ffecf2