New inference: deal with unions containing type variables in narrowing
Summary:
The `expand_type_and_narrow` function currently triggers narrowing of the type only on type variables. But frequently there is the situation where we have a union involving type variables and some concrete types e.g. `#1 | C`, with no useful bounds information on the type variable. In this situation, we will simply report an error e.g. "Was expecting an object but type is unknown", even though the concrete class in the union might be sufficient to suggest an upper bound on the whole type. For example, we are trying to access a property `$x->prop`, with `$x:#1|C` and the superclass of `C` is a class `B` that defines `prop`, it would make sense to require `(#1 | C) <: B` in order to type check the access. Contrast the case where `$x:#2` and `#1` and `C` are lower bounds on `#2`. In this situation we already impose `#2 <: B`, thereby accepting the property access expression.

So in this diff we deal with unions that contain type variables, treating them the same way as lower bounds on a type variable.

Reviewed By: manzyuk

Differential Revision: D14643216

fbshipit-source-id: 5ffa1e650fd06fc6e516c3f6c7afeb28f49570b3