Disable one more instance of memory sharing of little consequence
Summary:
This is similar to {D14574688}. The code is functionally the same, but this diff makes the memory layout more uniform - we always allocate new block for value, instead of the one-off case where we sometimes re-use the block of missing node. It makes byte-by-byte comparisons with Rust parser, where such memory sharing requires more intentionality, easier.

This memory sharing seems "accidental" anyway - it was likely a syntactic choice (the code before is slightly more compact), than a performance one.

Reviewed By: kmeht

Differential Revision: D14710205

fbshipit-source-id: f4d2657cbc5fa2fd025d19fbff9bd1ec08d4feb9