Reorder MRO to match Decl_inherit (again)
Summary:
In {D13579468}, I attempted to exactly reproduce in Decl_linearize the ordering implied by Decl_inherit, but I overlooked something.

In the original implementation of linearization, we reversed the order that traits were included in the linearization. This is because when multiple traits define a method of the same name, the one inherited by a class is the one defined in the trait which appears last syntactically in the class's `use` statements. In Decl_inherit, this behavior is implemented with a List.fold_left--since we copy all inherited trait methods into the declared class from left to right, multiple methods of the same name will be overwritten, and the one defined in the rightmost trait will be the one we keep. While working on this stack, I noticed that it was necessary to also reverse the list of XHP attribute inclusions, so I also made that change in Decl_linearize.

I do not know why the original implementation of Decl_linearize only reversed the list of traits, because [*every* fold over ancestors in Decl_inherit is a fold_left](diffusion/FBS/browse/master/fbcode/hphp/hack/src/decl/decl_inherit.ml;2f6437114a85aaf457f4dad981e04b2b4d3c28e2$434,437,441,444,451,453) except for [the one over parents, which is a fold_right](diffusion/FBS/browse/master/fbcode/hphp/hack/src/decl/decl_inherit.ml;6a9273e9040a28bfd4f86fe542ac1ff9fc4afc77$399) (again, perhaps the current behavior was chosen in order to match HHVM's behavior, but for now I want to model the typechecker's existing behavior as closely as possible). Thus, to reproduce the behavior of Decl_inherit, Decl_linearize should reverse every one of these lists except for the list of parents.

Reviewed By: arxanas

Differential Revision: D14738719

fbshipit-source-id: 8a85a188760425f60aabf3a7d8aa30b359093610