bump up retries
Summary:
If an hh invocation causes hh_server to spin up, then it takes a long time (a large number of once-per-second retries) before the server has spun up and is ready to serve requests.

We should make a proper solution where the retry count only ever refers to the time it takes to establish that there is a server, rather than the time it takes to connect to a working and running and responsive server. But that would be a major architectural change.

In the meantime, as a bandaid, I'm bumping up this default.

Reviewed By: 2BitSalute

Differential Revision: D14779164

fbshipit-source-id: 2322a56ed9246b880d96eeb84caba5474099ebf9