Suppress formatting of nodes following HH_FIXMEs only if they span 3 lines or fewer
Summary:
We have had a couple bug reports about hackfmt being a no-op on classes with HH_FIXMEs on the class definition. This is because HH_FIXMEs currently suppress formatting for the entire node following them. This guarantees that we don't accidentally move a suppressed error away from its suppression comment while formatting, but it is an overly aggressive and unintuitive policy.

It is probably possible to suppress formatting for only the line following the FIXME, but this would involve substantial and probably-risky changes which I do not have time for right now. Instead, we can use the number of newlines in the node as a simple heuristic for deciding whether it is reasonable to suppress formatting for it.

Reviewed By: arxanas

Differential Revision: D14786882

fbshipit-source-id: 26915c8bb4f3a12a893cfb8c8575adf2042b7dd2