Add a group for IssetL
Summary:
We were missing some optimization opportunities by not doing this. But
it can currently cause some correctness issues too. We currently
optimize `IssetL $this` to `BareThis;IsTypeC Null;Not`, but this gives
us better type inference than we had before the optimization, and as a
result we can see new optimizations; but since we don't (necessarily)
iterate in the optimize phase, we can make incorrect assertions
(analysis says we'll constprop something, but we don't run optimize
again, so we don't actually const-prop it).

This showed up when I added better type refinement for $this.

Reviewed By: ricklavoie

Differential Revision: D14766307

fbshipit-source-id: 841c97c7402260b484c3336a1cac8cf5b55aca10