Extend hh_server with a refactoring allowing to rewrite lambda parameters with suggested types
Summary:
The new type inference algorithm occasionally fails to infer the types of the parameters of a lambda if they don't have explicit type hints and the context in which the lambda is used doesn't provide enough type information.  This results in the new type checker generating errors for code that is actually safe (and that is often accepted by the legacy type checker).  One way to fix these errors is to annotate the parameters with type hints.  This diff introduces a new mode in `hh_server` that allows one to annotate the parameters of lambdas with the types suggested by the type checker.

Roughly, it works as follows:
- Suppose that we are checking a method call, like `$x->foo()`.  We need to know the type of the receiver `$x` to check if `foo` is defined for it.  If the type of `$x` is a type variable, we try to eagerly solve it, which typically means pushing the type variable to its lower bound (the union of the lower bounds stored in the type variable environment).  If the variable doesn't yet have lower bounds, which often happens if the receiver is a parameter of a lambda without an explicit type hint, like `$x ==> $x->foo()`, we report an "unknown type" error, record the fact that the eager solve failed for this type variable, and move on.  If later we are again forced to solve the type variable and we discover that it does have some lower bounds, we bind it to the type that is the union of the lower bounds with a special reason annotation: `Rsolve_fail`.  Often this is the type that would also work for type checking `$x->foo()` above, so it's the type we suggest as a possible type hint.
- This diff adds a new flag `--rewrite-lambda-parameters` to `hh_client` that accepts a list of files.  For each file, we generate its TAST, find all lambda parameters whose types are annotated with `Rsolve_fail`, and if the type is user-denotable (e.g., not a union, "any", expression dependent type etc.) then we generate an `Insert` patch that inserts the textual representation of the type in front of the parameter.  These patches are then applied by `ClientCheck`.

I'm not sure this is not a one-off diff and whether we want to have it landed eventually.

Reviewed By: andrewjkennedy

Differential Revision: D14578562

fbshipit-source-id: bc9ef28956bc8bc1863dfd511a5f994b6803f1c7