Fix specified function call precedence by improving look-a-head
Summary:
When specified function calls was initially implemented it didn't take into account that things to left of the specified function call would have a precedence that would result in stealing the identify away from the function call.

This means that things like this: `f<T>($x);` worked fine, but `42 * f<T>($x);` or `! f<T>($x)` didn't parse correctly.

This becomes an even larger problem w/ the upcoming await precedence change and reified generics.

This diff improves the look-a-head to properly identify these cases at the cost of ... looking-a-head more.

Reviewed By: oulgen

Differential Revision: D14761813

fbshipit-source-id: a247bd598b0d550d3dcbfea4a29cf090e1b2df54