Allow users to specify part of a mangled closure name
Summary:
The HHJS transpiler emits a godawful number of closures--we'd like to profile
the output code but it's difficult to do with the default name-mangling; and it
seems like overkill to do our own closure conversion that names things more sensibly.

This adds support for a user attribute `__ClosureName` that inserts a
user-provided string into the mangled closure name.

Not sure if this is a kosher thing to do or not--feedback on this approach is
more than welcome.

Reviewed By: oulgen

Differential Revision: D13975389

fbshipit-source-id: b83da6beb570bbc94b2a723d7904daf9cfadcf8c