Convert emit_lval_op() to Scope.with_unnamed_local()
Summary:
`emit_lval_op()` used `stash_in_local_with_prefix()` to implement list
assignment.

It was buggy. If RHS was a local that was used inside the list assignment on
LHS, it would be used directly instead of being stashed into unnamed local and
it would get overwritten in the middle of list unpacking. This bug does not
seem to be present in older versions of HHVM, so perhaps it was introduced by
HackC. It was also introduced by PHP5.1.0 and fixed by PHP7.0.0:
https://3v4l.org/SulFW

This bug was also causing behavioral difference between the typechecker and
runtime, as demonstrated by the newly added test (it typechecks just fine, but
fails on typecheck at runtime).

This diff converts usage of unnamed local to `Scope.with_unnamed_local()` and
does not copy the buggy optimization from `stash_in_local_with_prefix()`.
Instead, the optimization is left to hhbbc, which properly tracks local usage
and optimizes away locals only when it is safe to.

Reviewed By: alexeyt

Differential Revision: D14635269

fbshipit-source-id: 36b8a3860d3c26bb24eec06ecb61895da4108277