Fix use-after-scope in cli-server.cpp
Summary: Destructors are run in inverse order to their creation, so there's a small window where `wrapper` is destroyed but can still be used. This was causing my ASAN builds to crash with a use-after-scope error. The fix is to ensure the pointer to `wrapper` is unset before the `wrapper` object is actually destroyed, rather than shortly after.

Reviewed By: ricklavoie

Differential Revision: D14815979

fbshipit-source-id: db4435be03825e3ade03a21b6b90971feb26ca1a