Simplify unwinder, remove stack of Faults
Summary:
Removal of fault funclets and Catch opcode trivialized usage of the stack of
faults. This stack is always empty except for the duration of invocation of
unwindPhp(), when it stores the fault that is being processed by the unwinder.
Let's just use a variable for that and simplify the code.

Reviewed By: ricklavoie

Differential Revision: D14679006

fbshipit-source-id: 534bd6406312c8821d1b714cef3f4c9f386bdcd5