Unify list of autoimported functions between typechecker and compiler
Summary:
The typechecker and compiler have completely separate lists of which functions are autoimported. This diff unifies the two lists (pretty much just using the compiler's list, since it's more comprehensive).

Temporarily, we expose the list from `Namespaces` so `Hhbc_id` can use it, but a followup diff will unify the autoimporting implementations as well, allowing us to hide it as an implementation detail again. (Unifying the implementations is necessary to eliminate namespace fallback entirely, since the compiler's implementation of autoimport depends on fallback.)

Reviewed By: vassilmladenov

Differential Revision: D14841681

fbshipit-source-id: c6fac4137f075d1fe419bf195f27bc2fef056e14