Preserve backtrace when re-raising exceptions
Summary:
As far as I can tell, the code pattern
```
| e ->
  ...
  raise e
```
will raise an exception whose backtrace relates to the "raise" statement -- not the original backtrace of the code that first raised the exception.

I ran into this when investigating T42826658.

In this diff I did a BigGrep for all occurrences of "raise e", and made sure they all had the right backtraces. (I skipped a few that didn't seem very significant).

Is this the correct code pattern? I read through the related github issue tracker about "reraise" and came to the conclusion that the reraise external primitive was kind of hard to trust, and I'd much rather reraise using the explicitly provided primitives.

Differential Revision: D14862759

fbshipit-source-id: cfc329c7aac19b6b7d5275e6b3f19391589f4c56