New inference: narrow types for array access through a type variable
Summary:
Array access, assignment and append operations are overloaded, so in `Typing_array_access.ml` we do explicit case analysis over the type of the array. When the type is a type variable, we attempt to resolve it by `expand_type_and_solve` but sometimes this does not resolve the type.

Instead, for access and append operations for now, we can use `expand_type_and_narrow` to use the lower bounds of the type variable as a heuristic for "narrowing" the type by providing an upper bound. For example, if we know `vec<string> <: #1` then we can add a bound `vec<string> <: #1 <: vec<#2>` i.e. assume that the array some kind of `vec`. (Note that this closes off the possibility of adding another lower bound such as `Vector<string>` that might be revealed when checking more code.)

We also apply narrowing to the refinement of shapes.

Reviewed By: CatherineGasnier

Differential Revision: D14666297

fbshipit-source-id: 0e8d82796cd9695250d0522ad065eaf016e805c0