Add comment about potential bug
Summary:
I saw a bug report with this timeline:
```
2:52:42 - textDocument/didOpen "DeactivateYourAccountParams.php"
2:53:35 - old hh_client LSP server terminates and a new one starts
2:55:15 - textDocument/didClose "DeactivateYourAccountParams.php"
2:55:18 - error relating to didClose
```

There seemed to be an error handling the didClose, on the grounds that serverFileSync.ml::close_relative_path was invoked on the second instance of hh_server, and this code expects to see that the file had been previously opened by this instance, but I guess it wasn't.

I haven't been able to repro in any controlled deterministic way. I know the problem will in any case go away with the ServerlessIDE rewrite. Just wanted to write something down in the code for posterity.

Reviewed By: arxanas

Differential Revision: D14864488

fbshipit-source-id: 87123b84f40e47d228fe5b1677fccb15f379da2a