Don't add newline before semicolon in hhi-style function signatures
Summary:
We added a newline here in order to handle loops without a compound statement (which use the same logic that we use to handle function bodies):

```
while (true)
  do_something();
```

This is not the right policy for function signatures like the added test case, where the "function body" node is just a semicolon token. If the "function body" is just a Token (rather than a CompoundStatement node, as usual, or a Statement node, as in the above example), then don't add a newline or indentation.

Reviewed By: arxanas

Differential Revision: D14875373

fbshipit-source-id: f0c23d26054bb633005c0676ee7df8d6999585d0