Fix an issue with interface func families
Summary:
For interfaces, we define FuncFamilies for each method that is
guaranteed to be implemented on all classes that implement the
interface. This includes methods that are not actually declared by the
interface.

For abstract classes, we only define FuncFamilies for methods actually
declared on the class.

This can cause a problem if early on we deduce that a variable is a
subtype of an interface, which has a well defined funcfamily, but
later on we find that the variable is actually a subtype of an
abstract class that implements that interface, but doesn't explicitly
declare the method in question.

In that case, although we now know more about the type, we know less
about the type returned by the method. The new test case hits this
issue, and asserts because index types become less refined.

This refactors the code a bit, to use an interface's subclassList in
more or less the same way as other classes (the difference being that
we only include instantiable classes on it). This gets rid of the need
for ifaceImplementerMap, and makes it possible to share
define_func_family between normal classes, and interfaces.

Reviewed By: paulbiss

Differential Revision: D14853752

fbshipit-source-id: b3ab3fa2806e38bda00eb24bb2efa114d9accc16