Fix review.sh
Summary:
Looks like the change to review.sh in {D14854837} had an error--the variable `$SRC` is not defined. This causes review.sh to fail to copy the actual output into the expected output file when requested--instead, it copies the actual output into a file named ".exp".

I think the intent of the change in D14854837 was to always copy into the file with the EXP_EXT extension rather than FALLBACK_EXP_EXT. If that's the behavior we want, then all we need to do then is replace `$SRC` with the actual filename, `$f`.

Reviewed By: kmeht

Differential Revision: D14883378

fbshipit-source-id: 99ec1af742bb6291d74ff0f04f4069f75ad23db5