New native HH\int_mul_overflow, HH\int_mul_add_overflow
Summary:
Adding a pair of functions to get the overflow part of a 64-bit multiplication (unsigned), so that a very fast and general version of fastrange64 can be implemented and used in Hack code. See https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/

The second function also takes a 'bias' parameter that is added before returning the overflow result. This should enable implementing a variant of fastrange64 whose weighted average return value is always exactly equal to the unweighted average. This would be a useful feature in experimentation, such as converting to a 0.0 to 1.0 float without bias in simple cases like 50/50 A/B tests.

For example (suppose we were doing 4 bit overflow instead of 64 bit):
  int4_mul_overflow($x, 10)
    0 -> 0
    1 -> 0
    2 -> 1
    3 -> 1  (30 div 16)
    4 -> 2  (40 div 16)
    5 -> 3  (etc.)
    6 -> 3
    7 -> 4
    8 -> 5
    9 -> 5
   10 -> 6
   11 -> 6
   12 -> 7
   13 -> 8
   14 -> 8
   15 -> 9
a.k.a. weights are
  2,2,1,2,1,  2,2,1,2,1
but if we correct by 10/2 === 5 we get this:
  int4_mul_add_overflow($x, 10, 5)
    0 -> 0
    1 -> 0
    2 -> 1
    3 -> 2  (35 div 16)
    4 -> 2  (45 div 16)
    5 -> 3  (etc.)
    6 -> 4
    7 -> 4
    8 -> 5
    9 -> 5
   10 -> 6
   11 -> 7
   12 -> 7
   13 -> 8
   14 -> 9
   15 -> 9
a.k.a. weights are
  2,1,2,1,2,  2,1,2,1,2
Which is mirrored bottom-to-top (and within every subdivision by a power-of-2, here just 2) thus weighted average is equal to unweighted average.

What about an odd range?

  int4_mul_overflow($x, 7)
    0 -> 0
    1 -> 0
    2 -> 0
    3 -> 1  (21 div 16)
    4 -> 1  (etc.)
    5 -> 2
    6 -> 2
    7 -> 3
    8 -> 3
    9 -> 3
   10 -> 4
   11 -> 4
   12 -> 5
   13 -> 5
   14 -> 6
   15 -> 6
a.k.a. weights are
  3,2,2,3,2,2,2
With a bias:
  int4_mul_add_overflow($x, 7, 3 or 4)
    0 -> 0
    1 -> 0
    2 -> 1  (17 or 18 div 16)
    3 -> 1  (etc.)
    4 -> 1
    5 -> 2
    6 -> 2
    7 -> 3
    8 -> 3
    9 -> 4
   10 -> 4
   11 -> 5
   12 -> 5
   13 -> 5
   14 -> 6
   15 -> 6
a.k.a. weights are
  2,3,2,2,2,3,2
Testing with range = 5 indicates that the floor of range / 2 should be used.

Reviewed By: markw65

Differential Revision: D14839003

fbshipit-source-id: 4815c952229f4bb4157071f18efbd98d228bf34a