Fix a potential issue with public static props
Summary:
When we resolve an Init[S]Prop we update the class's initial value,
but we didn't update the initial value in
ClassInfo::publicStaticProps. If we no longer see any mutations for
that property, it will get a type of TBottom.

Currently this isn't a problem, because we only modify the bytecode
during the constant pass, and the final optimization pass, and we
don't start tracking public statics until after the constant pass. A
followup diff changes that to update the bytecode after every analysis
pass, which exposes the issue.

Reviewed By: ricklavoie

Differential Revision: D14854331

fbshipit-source-id: 9494c3ccd7ba6aa37cf87442ef877f7130b6c11a